Don't try delete the message unless we can

+++ command_handlers/gdpr_consent.js
@@ -5,7 +5,9 @@ module.exports = {
   handler: async function(message) {
     await db.pool.query('UPDATE users SET gdpr_consented=true WHERE discord_id=$1', [message.author.id]);
     await message.author.send('Your consent preference has been updated. Thank you!');
-    await message.delete();
+    if (message.channel.type == 'text') {
+      await message.delete();
+    }
   },
   help: 'Allow the bot to store and process your personal information',
 };
